added task delete and update functionility

+++ src/dao/TaskRepo.java
@@ -16,9 +16,29 @@ public class TaskRepo {
         return taskDB;
     }
 
-    public void removeTask(int id){
+    public boolean removeTask(int id){
 
-        taskDB.removeIf(task -> task.getId() == id);
+        if(id>taskDB.size()-1)
+        {
+            System.out.println("Enter the valid taskId");
+            return false;
+        }
 
+        return taskDB.removeIf(task -> task.getId() == id);
+
+    }
+
+    public boolean updateTask(int id){
+
+        if(id>taskDB.size()-1){
+            System.out.println("Task does not exists!");
+            return false;
+        }
+
+        return true;
+    }
+
+    public Task getTask(int id){
+        return taskDB.get(id);
     }
 }
